moved compiler iterators to separate file; custom file position iterator
the utf8 to utf32 iterator returns a value instead of a reference, violating the ForwardIterator requirements that boost::spirit::classic::position_iterator relies on. (This results in returning a reference to a local variable, which does not work.) My custom iterator works around that and adheres to the ForwardIterator Concept as much as possible.