Started grammar implementation
I'm mostly fighting boost.spirit at the moment; getting the types quite right is a long struggle. The documentation is a little spotty, some things are never quite explained (like boost::spirit::qi::token) and existing documentation is often hard to find due to the structure... But if it compiles it's probably right, yay templates!