Split AST into parser & clean version
The clean variant has proper binary operations, which are impossible for the parser to generate due to left recursion.

This incurs quite a bit of duplication and some busy work when converting, but this separation keeps the number of expression variants low for each kind of AST, preventing dummy visitor entries for unused node types.