Either around bug or error detection.
I hate this, but I tried to figure it out for hours and I can neither reproduce the problem nor explain it, so I'm using what you may consider a hack. It's all legal code, mind; I'm using raw pointers instead of iterators to refer to registered functions, which the standard guarantees to keep their address when further insertions happen.

It's just that the previous iterator-based solution should have worked just as nicely. It crashed during iterator destruction, presumably due to invalidation, which I for the life of me could not comprehend. I'm blaming it on some bug in the interplay of boost::variant and the msvc iterators, but here's the ugly part: I got the same exceptions legitimately for the error fixed in [90da13e], so chances are there is actually an error on my part and I'm not seeing it.

So basically, I'm hoping to look into this later, but for now it seems this hacky solution appears to work, and I need some quick results right now. (Famous last words, eh?) I'm branching this in hopes of keeping the master branch clean.